fix: minor improvements for merge
