+++ app/api/sep1/route.ts
@@ -1,0 +1,102 @@
+import { retrieveOrganization } from "@/actions/organization";
+import { retrieveActiveProductsWithAsset } from "@/actions/product";
+import TOML from "@iarna/toml";
+
+const DEFAULT_LOGO_URL = "https://stellartools.io/default-logo.png";
+const PLATFORM_WEB_AUTH_ENDPOINT = "https://api.stellartools.io/auth";
+const PLATFORM_PHYSICAL_ADDRESS =
+  "123 Stellar Way, San Francisco, CA 94102, United States";
+const PLATFORM_OFFICIAL_EMAIL = "support@stellartools.io";
+const TESTNET_PASSPHRASE = "Test SDF Network ; September 2015";
+const MAINNET_PASSPHRASE = "Public Global Stellar Network ; September 2015";
+
+export async function GET(request: Request) {
+  const host = request.headers.get("host") || "";
+  const orgSlug = host.split(".")[0];
+
+  if (!orgSlug) {
+    return new Response("Organization not found", { status: 404 });
+  }
+
+  try {
+    const org = await retrieveOrganization({ slug: orgSlug });
+    const stellarAccount = org.stellarAccounts?.[org.environment];
+
+    const productsWithAssets = await retrieveActiveProductsWithAsset(
+      org.id,
+      org.environment
+    );
+
+    const networkPassphrase =
+      org.environment === "testnet" ? TESTNET_PASSPHRASE : MAINNET_PASSPHRASE;
+
+    const orgLogo = org.logoUrl || DEFAULT_LOGO_URL;
+    const orgUrl = `https://${orgSlug}.stellartools.io`;
+    const orgSupportEmail = `${orgSlug}@stellartools.io`;
+    const orgDescription =
+      org.description || `${org.name} - Powered by Stellar Tools`;
+
+    const currencies = productsWithAssets.map(({ product, asset }) => {
+      const issuer = asset.issuer;
+      const productImage = product.images?.[0] || org.logoUrl;
+
+      const currency: Record<string, unknown> = {
+        code: asset.code,
+        display_decimals: 7,
+        name: product.name,
+        desc: product.description || `${product.name} by ${org.name}`,
+        image: productImage || orgLogo,
+        status: "live",
+        is_asset_anchored: false,
+        redemption_instructions: `Purchase at ${orgUrl}/checkout`,
+      };
+
+      if (issuer) currency.issuer = issuer;
+
+      if (product.billingType === "metered") {
+        currency.conditions = `Metered billing: ${product.creditsGranted} credits included, expires in ${product.creditExpiryDays} days`;
+      } else if (product.billingType === "recurring") {
+        currency.conditions = "Recurring subscription";
+      }
+
+      return currency;
+    });
+
+    const tomlData = {
+      NETWORK_PASSPHRASE: networkPassphrase,
+      ACCOUNTS: stellarAccount?.public_key ? [stellarAccount.public_key] : [],
+      WEB_AUTH_ENDPOINT: PLATFORM_WEB_AUTH_ENDPOINT,
+      DOCUMENTATION: {
+        ORG_NAME: org.name,
+        ORG_URL: orgUrl,
+        ORG_LOGO: orgLogo,
+        ORG_DESCRIPTION: orgDescription,
+        ORG_PHYSICAL_ADDRESS: PLATFORM_PHYSICAL_ADDRESS,
+        ORG_OFFICIAL_EMAIL: PLATFORM_OFFICIAL_EMAIL,
+        ORG_SUPPORT_EMAIL: orgSupportEmail,
+      },
+      PRINCIPALS: [
+        { name: "Stellar Tools Support", email: PLATFORM_OFFICIAL_EMAIL },
+      ],
+      CURRENCIES: currencies,
+    };
+
+    const tomlString = TOML.stringify(
+      tomlData as Parameters<typeof TOML.stringify>[0]
+    );
+
+    return new Response(tomlString, {
+      headers: {
+        "Content-Type": "text/plain; charset=utf-8",
+        "Access-Control-Allow-Origin": "*",
+        "Cache-Control": "public, max-age=3600",
+      },
+    });
+  } catch (error) {
+    if (error instanceof Error && error.message === "Organization not found") {
+      return new Response("Organization not found", { status: 404 });
+    }
+
+    return new Response("Internal server error", { status: 500 });
+  }
+}
